feat: added configurable dtype to to_torch util

+++ cardio_rl/utils.py
@@ -11,5 +11,5 @@ def to_jnp(arr: np.ndarray, dtype=jnp.float32):
     return jnp.asarray(arr, dtype=dtype)
 
 
-def to_torch(arr: np.ndarray):
-    return th.from_numpy(np.array(arr)).float()
+def to_torch(arr: np.ndarray, dtype=th.float32):
+    return th.from_numpy(np.array(arr)).to(dtype)
